+++ frontend/public/changelog.md
@@ -4,6 +4,7 @@
 
 - Menampilkan ragam varian dan kiasan di redaksi kamus
 - Menambahkan kategori akronim dan singkatan di bentuk bebas serta kiasan di ekspresi
+- Memproses tampilan entri glosarium agar mengenali titik koma dan tanda kurung
 
 ## 2026-02-23 Senin